Remove protobuf depenency
Original motivation for this change was to
remove unnecessary binary dependency on
libatomic. As it appeared source of this dependency
was protobuf. As this quite complex piece of
software was used only to generate simple json
I swiched it to boost::json.
I decided to use boost::json because boost already was
dependency for eBPF discovery.