Read all interfaces (IPv4 and IPv6) from getifaddrs (#111)
Unification of detecting if address is external for IPv4 and IPv6. All interfaces (IPv4 and IPv6) are read from getifaddrs. 
Removed special handling of bridges in IPv4. Address is detected as external if it does not belong to reserved private range and does not belong to network of local interfaces.